requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-mysql==3.12.0
djangorestframework-camel-case==1.2.0
mysqlclient==2.0.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-rest-auth==0.9.5
django-imagekit==4.0.2
django-environ==0.9.0
environ==1.0
pillow
pandas
numpy
requests~=2.26.0
gunicorn==20.1.0